+++ src/main/resources/policy-account.yml
@@ -1,10 +1,13 @@
 business:
   rule:
     account:
-      # 일일 이체 한도 (원)
-      main-daily-limit: 3000000
-      # 자동 충전 단위 (원)
-      charge-unit: 10000
-    saving-account:
-      fixed-interest-rate: 0.05
-      flexible-interest-rate: 0.03
+      main:
+        # 일일 이체 한도 (원)
+        main-daily-limit: 3000000
+        # 자동 충전 단위 (원)
+        charge-unit: 10000
+        account-prefix: "01"
+      saving:
+        fixed-interest-rate: 0.05
+        flexible-interest-rate: 0.03
+        account-prefix: "02"
